ci: jailer-lite-property + firecracker-bridge-fuzz lanes (Plan 113 §Task 15)
Two new CI lanes plus the refactor that makes the fuzz lane feasible:

jailer-lite-property — runs Task 9's #[ignore]-gated seccomp +
  Landlock property tests on Ubuntu 22.04 (kernel >= 5.19) on every
  PR. Skips on runners without Landlock sysfs node.

firecracker-bridge-fuzz — adversarial fuzz of the BridgeConfigJson
  serde_json parser and the PasstHashesFile toml parser. Folded into
  security.yml's existing fuzz job (manual dispatch + nightly +
  release-tag), mirroring fuzz_supervisor_config / fuzz_guest_request
  / unpack_layer.

Refactor: mvm-firecracker-bridge gains a lib.rs + src/parse.rs
exposing BridgeConfigJson, PasstHashesFile, and verify_passt_hash
as public items. main.rs is unchanged at the call-site level; this
just makes the parsers reachable for fuzz harness use.

Plan: specs/plans/113-network-provider-trait-firecracker-substrate.md §Task 15.
